# Bytes trits conversion