--- a/docs/html/doxygen.css
+++ b/docs/html/doxygen.css
@@ -1,4 +1,4 @@
-/* The standard CSS for doxygen 1.8.13 */
+/* The standard CSS for doxygen 1.8.14 */
 
 body, table, div, p, dl {
 	font: 400 14px/22px Roboto,sans-serif;