--- a/docs/html/search/all_11.js
+++ b/docs/html/search/all_11.js
@@ -1,5 +1,5 @@
 var searchData=
 [
   ['validatesignatures',['validateSignatures',['../namespace_i_o_t_a_1_1_crypto_1_1_signing.html#a86f01ade2568237125f08511d6aa85bb',1,'IOTA::Crypto::Signing']]],
-  ['verifybundle',['verifyBundle',['../class_i_o_t_a_1_1_a_p_i_1_1_extended.html#a2ec2d0dfeb48ffccdc4d526e2e137729',1,'IOTA::API::Extended']]]
+  ['verifybundle',['verifyBundle',['../class_i_o_t_a_1_1_a_p_i_1_1_extended.html#ae24ceda0617d2c93d279540f603d809c',1,'IOTA::API::Extended']]]
 ];