--- a/docs/html/search/classes_1.js
+++ b/docs/html/search/classes_1.js
@@ -2,6 +2,7 @@ var searchData=
 [
   ['badrequest',['BadRequest',['../class_i_o_t_a_1_1_errors_1_1_bad_request.html',1,'IOTA::Errors']]],
   ['base',['Base',['../class_i_o_t_a_1_1_a_p_i_1_1_responses_1_1_base.html',1,'IOTA::API::Responses::Base'],['../class_i_o_t_a_1_1_a_p_i_1_1_requests_1_1_base.html',1,'IOTA::API::Requests::Base']]],
+  ['bigint',['Bigint',['../class_i_o_t_a_1_1_types_1_1_bigint.html',1,'IOTA::Types']]],
   ['broadcasttransactions',['BroadcastTransactions',['../class_i_o_t_a_1_1_a_p_i_1_1_requests_1_1_broadcast_transactions.html',1,'IOTA::API::Requests']]],
   ['bundle',['Bundle',['../class_i_o_t_a_1_1_models_1_1_bundle.html',1,'IOTA::Models']]]
 ];