--- a/docs/html/search/functions_1.js
+++ b/docs/html/search/functions_1.js
@@ -2,8 +2,9 @@ var searchData=
 [
   ['badrequest',['BadRequest',['../class_i_o_t_a_1_1_errors_1_1_bad_request.html#ace1e22b3f4e62de18332d7d67fd8ed4e',1,'IOTA::Errors::BadRequest']]],
   ['base',['Base',['../class_i_o_t_a_1_1_a_p_i_1_1_requests_1_1_base.html#aba5d298949aa7a6761e6cbefd0181676',1,'IOTA::API::Requests::Base::Base()'],['../class_i_o_t_a_1_1_a_p_i_1_1_responses_1_1_base.html#ac5f64bc6a926f8419985c71e5bf16716',1,'IOTA::API::Responses::Base::Base(const int64_t &amp;duration=0)'],['../class_i_o_t_a_1_1_a_p_i_1_1_responses_1_1_base.html#a37b5339e7c74153f0b1ceffac3af7f46',1,'IOTA::API::Responses::Base::Base(const json &amp;res)']]],
+  ['bigint',['Bigint',['../class_i_o_t_a_1_1_types_1_1_bigint.html#aa60058f8209380f7441d6f709bcd991d',1,'IOTA::Types::Bigint']]],
   ['broadcastandstore',['broadcastAndStore',['../class_i_o_t_a_1_1_a_p_i_1_1_extended.html#aa84894cbaa3fa57252a8a34a75e0d7ab',1,'IOTA::API::Extended']]],
   ['broadcasttransactions',['broadcastTransactions',['../class_i_o_t_a_1_1_a_p_i_1_1_core.html#a0f9605f23dae0920d286d208a5ad806e',1,'IOTA::API::Core::broadcastTransactions()'],['../class_i_o_t_a_1_1_a_p_i_1_1_requests_1_1_broadcast_transactions.html#a24d06ea29f9ba55533d4e2f5bce6dd06',1,'IOTA::API::Requests::BroadcastTransactions::BroadcastTransactions()']]],
   ['bundle',['Bundle',['../class_i_o_t_a_1_1_models_1_1_bundle.html#a37fba389b7926a81454473e8ae6151a6',1,'IOTA::Models::Bundle::Bundle()=default'],['../class_i_o_t_a_1_1_models_1_1_bundle.html#aaa83c4c0a0ee722aa5bbc44f2fc9c9a9',1,'IOTA::Models::Bundle::Bundle(const std::vector&lt; Models::Transaction &gt; &amp;transactions)']]],
-  ['bundlesfromaddresses',['bundlesFromAddresses',['../class_i_o_t_a_1_1_a_p_i_1_1_extended.html#a43c72cf1f516d9f10c23b3ac3cb01899',1,'IOTA::API::Extended']]]
+  ['bundlesfromaddresses',['bundlesFromAddresses',['../class_i_o_t_a_1_1_a_p_i_1_1_extended.html#ab53e93d182ec0cf92eef6831ffbb1539',1,'IOTA::API::Extended']]]
 ];