--- a/include/iota/types/big_int.hpp
+++ b/include/iota/types/big_int.hpp
@@ -72,14 +72,15 @@ public:
   std::vector<int8_t> toBytes() const;
 
 private:
-  bool         is_negative() const;
-  uint32_t     mul(uint8_t factor, unsigned int ms_index);
-  uint32_t     div(uint8_t divisor);
-  bool         add(const uint32_t *a, const uint32_t *b);
-  bool         sub(const uint32_t *a, const uint32_t *b);
-  unsigned int add_u32(uint32_t summand);
-  int          cmp(const uint32_t *b) const;
-  bool         set_last_trit_zero();
+  inline bool         isNegative() const;
+  inline uint32_t     mul(uint8_t factor, unsigned int ms_index);
+  inline uint32_t     div(uint8_t divisor);
+  inline bool         add(const uint32_t *a, const uint32_t *b);
+  inline bool         sub(const uint32_t *a, const uint32_t *b);
+  inline unsigned int addU32(uint32_t summand);
+  inline bool         addcarryU32(uint32_t *r, uint32_t a, uint32_t b, bool c_in);
+  inline int          cmp(const uint32_t *b) const;
+  inline bool         setLastTritZero();
 
 private:
   uint32_t data[WordHashLength];