Do not merge yet, triggering travis.

Still need to :

- [x] Write documentation
- [x] Rewrite tests
- [x] Do some benchmarking
- [x] Clean the code