@Cylix 2018-03-10
Benchmarks (tested on a debian 9 running on VMWare configured with 4 core and 4GB RAM, compiled with -O3 and not debug flags):

* iota.lib.cpp before: 1.3s < t < 1.4s
* iota.lib.cpp after: 1.1s < t < 1.2s
* c-light-wallet: TBD

Also here is the evolution of pprof report:

* Before:
<img width="746" alt="screen shot 2018-03-09 at 11 21 08 pm" src="https://user-images.githubusercontent.com/3252462/37239445-23636e0c-23f1-11e8-936e-e821c155cdf9.png">


* After:
<img width="1148" alt="screen shot 2018-03-09 at 11 23 34 pm" src="https://user-images.githubusercontent.com/3252462/37239448-2779372e-23f1-11e8-9dff-446c2357e34d.png">