@Cylix 2018-03-11
Just tried the c library, it does the generation in ~0.1s, so roughly 10 times faster than us, we need to dig more.