@Cylix 2018-03-11
interestingly, while our program spend most of its time doing bigint::div, running pprof on c library does not even report anything related to division. We need to understand better the flow of both libraries.

<img width="522" alt="screen shot 2018-03-10 at 6 41 41 pm" src="https://user-images.githubusercontent.com/3252462/37248983-bcc843be-2492-11e8-9c52-f850299d3d51.png">

[output.pdf](https://github.com/thibault-martinez/iota.lib.cpp/files/1799989/output.2.pdf)