@Cylix 2018-03-11
Edit: this is executed, but the program executes so fast that the sample rate is not enough to capture the execution of that kind of small function (only captured 10 samples and when I increased a loooot the sample rate, I can get at most 30 samples that do not even show anything related to mul AND div).

I'm wondering if the compiler is more performant at optimizing the c div/mul functions rather than the cpp ones? Would need to check the generated assembly..