@Cylix 2018-03-11
c lib div function, no optimization (putting here for myself)

	.p2align	4, 0x90
_bigint_div_byte_mem:                   ## @bigint_div_byte_mem
	.cfi_startproc
## BB#0:
	pushq	%rbp
Lcfi72:
	.cfi_def_cfa_offset 16
Lcfi73:
	.cfi_offset %rbp, -16
	movq	%rsp, %rbp
Lcfi74:
	.cfi_def_cfa_register %rbp
	movb	%sil, %al
	movq	%rdi, -8(%rbp)
	movb	%al, -9(%rbp)
	movl	$0, -16(%rbp)
	movl	$12, -20(%rbp)
LBB24_1:                                ## =>This Inner Loop Header: Depth=1
	movl	-20(%rbp), %eax
	movl	%eax, %ecx
	addl	$-1, %ecx
	movl	%ecx, -20(%rbp)
	cmpl	$0, %eax
	jbe	LBB24_3
## BB#2:                                ##   in Loop: Header=BB24_1 Depth=1
	movl	$4294967295, %eax       ## imm = 0xFFFFFFFF
	movl	%eax, %ecx
	movl	-16(%rbp), %eax
	movl	%eax, %edx
	shlq	$32, %rdx
	movq	-8(%rbp), %rsi
	movl	-20(%rbp), %eax
	movl	%eax, %edi
	movl	(%rsi,%rdi,4), %eax
	movl	%eax, %esi
	addq	%rsi, %rdx
	movq	%rdx, -32(%rbp)
	movq	-32(%rbp), %rax
	movzbl	-9(%rbp), %r8d
	movl	%r8d, %edx
	xorl	%r8d, %r8d
	movq	%rdx, -40(%rbp)         ## 8-byte Spill
	movl	%r8d, %edx
	movq	-40(%rbp), %rsi         ## 8-byte Reload
	divq	%rsi
	movl	%edx, %r8d
	movl	%r8d, -16(%rbp)
	movq	-32(%rbp), %rdx
	movzbl	-9(%rbp), %r8d
	movl	%r8d, %edi
	movq	%rdx, %rax
	xorl	%r8d, %r8d
	movl	%r8d, %edx
	divq	%rdi
	andq	%rcx, %rax
	movl	%eax, %r8d
	movq	-8(%rbp), %rax
	movl	-20(%rbp), %r9d
	movl	%r9d, %ecx
	movl	%r8d, (%rax,%rcx,4)
	jmp	LBB24_1
LBB24_3:
	movl	-16(%rbp), %eax
	popq	%rbp
	retq
	.cfi_endproc