@Cylix 2018-03-11
cpp lib div function, no optimization (putting here for myself)

```
	.globl	__ZN4IOTA5Types6Bigint3divEh
	.p2align	4, 0x90
__ZN4IOTA5Types6Bigint3divEh:           ## @_ZN4IOTA5Types6Bigint3divEh
	.cfi_startproc
## BB#0:
	pushq	%rbp
Lcfi63:
	.cfi_def_cfa_offset 16
Lcfi64:
	.cfi_offset %rbp, -16
	movq	%rsp, %rbp
Lcfi65:
	.cfi_def_cfa_register %rbp
	movb	%sil, %al
	movq	%rdi, -8(%rbp)
	movb	%al, -9(%rbp)
	movq	-8(%rbp), %rdi
	movl	$0, -16(%rbp)
	movl	$12, -20(%rbp)
	movq	%rdi, -40(%rbp)         ## 8-byte Spill
LBB21_1:                                ## =>This Inner Loop Header: Depth=1
	movl	-20(%rbp), %eax
	movl	%eax, %ecx
	addl	$-1, %ecx
	movl	%ecx, -20(%rbp)
	cmpl	$0, %eax
	jbe	LBB21_3
## BB#2:                                ##   in Loop: Header=BB21_1 Depth=1
	movl	$4294967295, %eax       ## imm = 0xFFFFFFFF
	movl	%eax, %ecx
	movl	-16(%rbp), %eax
	movl	%eax, %edx
	shlq	$32, %rdx
	movl	-20(%rbp), %eax
	movl	%eax, %esi
	movq	-40(%rbp), %rdi         ## 8-byte Reload
	movl	8(%rdi,%rsi,4), %eax
	movl	%eax, %esi
	addq	%rsi, %rdx
	movq	%rdx, -32(%rbp)
	movq	-32(%rbp), %rax
	movzbl	-9(%rbp), %r8d
	movl	%r8d, %edx
	xorl	%r8d, %r8d
	movq	%rdx, -48(%rbp)         ## 8-byte Spill
	movl	%r8d, %edx
	movq	-48(%rbp), %rsi         ## 8-byte Reload
	divq	%rsi
	movl	%edx, %r8d
	movl	%r8d, -16(%rbp)
	movq	-32(%rbp), %rdx
	movzbl	-9(%rbp), %r8d
	movl	%r8d, %r9d
	movq	%rdx, %rax
	xorl	%r8d, %r8d
	movl	%r8d, %edx
	divq	%r9
	andq	%rcx, %rax
	movl	%eax, %r8d
	movl	-20(%rbp), %r10d
	movl	%r10d, %eax
	movl	%r8d, 8(%rdi,%rax,4)
	jmp	LBB21_1
LBB21_3:
	movl	-16(%rbp), %eax
	popq	%rbp
	retq
	.cfi_endproc
```