@Cylix 2018-03-11
The -O0 does not seem to differenciate much (2/3 more move instructions on c++, but that's negligible).

The -O3 is more interesting though.

The cpp version gives the following:

```
	.globl	__ZN4IOTA5Types6Bigint3divEh
	.p2align	4, 0x90
__ZN4IOTA5Types6Bigint3divEh:           ## @_ZN4IOTA5Types6Bigint3divEh
	.cfi_startproc
## BB#0:
	pushq	%rbp
Lcfi54:
	.cfi_def_cfa_offset 16
Lcfi55:
	.cfi_offset %rbp, -16
	movq	%rsp, %rbp
Lcfi56:
	.cfi_def_cfa_register %rbp
                                        ## kill: %ESI<def> %ESI<kill> %RSI<def>
	xorl	%edx, %edx
	movl	$13, %ecx
	.p2align	4, 0x90
LBB15_1:                                ## =>This Inner Loop Header: Depth=1
	shlq	$32, %rdx
	movl	(%rdi,%rcx,4), %eax
	orq	%rdx, %rax
	xorl	%edx, %edx
	divq	%rsi
	movl	%eax, (%rdi,%rcx,4)
	shlq	$32, %rdx
	movl	-4(%rdi,%rcx,4), %eax
	orq	%rdx, %rax
	xorl	%edx, %edx
	divq	%rsi
	movl	%eax, -4(%rdi,%rcx,4)
	addq	$-2, %rcx
	cmpq	$1, %rcx
	jne	LBB15_1
## BB#2:
	movl	%edx, %eax
	popq	%rbp
	retq
	.cfi_endproc
```

Which is slightly improved version of the -O0 version, dropping lots of move operation.
However, the two divq instructions are still there.

For the C version (not pasted here), the code is much more complex to go through, mostly due to lots of inlining.
The weird thing however is the total absence of any div instructions in the optimized c assembly.
I'm not sure how is that possible.